Parsing `#` sign inserted by preprocessor in fixed Form Fortran
In Fortran a file that is included is replaced by the doxygen preprocessor by something like:
`#    11 "<path>/<filename>" 2`
when this is in fixed form Fortran and the `<path>/<filename>` is long it is possible that the closing double quote is beyond position 72 and thus not seen resulting in:
```
Error in file <path>/<filename> line: 73, state: 22(String)
```
this is caused as the `#` is not handled properly in prepassFixedForm. The `#` should be handled analogous to e.g. `C`.