HTML start comment with 3 dashes
A normal HTML comment `<!--` has 2 `-` signs but it is not prohibited to have 3, but doxygen translates `<!---` into `<!-&ndash;` and thus the comment is not recognized.
By checking and consequently handling, doing the right skipping, also the 3 `-` sign version the problem can be solved.
An end comment cannot contain 3 `-` signs, so here no changes have to take place.
The version with 3 or more `-` are in a start HTML comment does not give a problem with xmllint either.