Warning with xrefitem from documentation
The example code with the `xrefitem` in the documentation gives a warning:
```
my_errors:1: warning: multiple use of page label 'my_errors', (other occurrence: .../aa.h, line: 4)
```
it is possible (and explicitly stated) to have an `\page` with the same name giving some extra information.

The warning regarding the multiple page is removed in case of a xrefitem page.