Skip empty HTML tag
Due to the fact that an empty HTML tag (like `<strike/>` is seen as a start tag the rest of a documentation block is handled with the strikethrough. Same problem occurs for other tags as well, in most cases empty HTML tags should be left out of the output (and just give the warning about the empty tag).

Mentioned example:
```
/**
Before a strike:
<strike/>
after strike
*/
void fie_b_strike(void);
```