Warning with XML / HTML style commands
When having a problem like:
```
///    set <var> var start_interp </em>
void fie(void);

///    set <i> i start_interp
void fie1(void);
```
we get the warnings (doubles omitted):
```
.../bb.h:5: warning: end of comment block while expecting command </em>
```
so
- discrepancy between `<var>` and closing `</em>` is not reported, but wrong for as opening and closing tag should have the same tag name.
- the missing `</i>` is reported as a missing `</em>`

The problems here are due to the fact that `<var>`, `<i>` and `<em>` share all the style `HTML_EMPHASIS`, this problem has been fixed by adding the used tag name to the style information and testing o
n the name and not the style type.

The result is now:
```
.../bb.h:1: warning: found </em> tag while expecting </var>
.../bb.h:2: warning: end of comment block while expecting command </var>
.../bb.h:5: warning: end of comment block while expecting command </i>
```
so now the real problems are reported.