Better warning in case a graph would have been to large
When a graph is to large we get a message like:
```
warning: Included by graph for 'tile.h' not generated, too many nodes. Consider increasing DOT_GRAPH_MAX_NODES.
```
but the user has no idea how big the graph really is so the value is added to the warning, so the user can decide to set a value larger than this value.