Not properly recognizing text in single quotes.
When we have a problem like:
```
   'http://search.cpan.org/search'

   "http://search.cpan.org/search"
```

The first line will be shown as: `'http://search.cpan.org/search'` where though part `http://search.cpan.org/search'` is hyperlinked (including the closing single quote!) resulting in a non-existing link. In case of the double quotes we just get `"http://search.cpan.org/search"` without links.

In the patch the single and double quote are handled in the same way