Length of identifying fenced code block
In case we have a problem like:
```
/** \file
 * ```
 *  router = (new Router());
 * ```
 */
```
there is no problem, but as soon as we add a space somewhere before the start or end identifier of the fenced code block we get:
```
.../aa.h:6: warning: reached end of file while inside a '```' block!
The command that should end the block seems to be missing!
```
i.e. a problem like:
```
/** \file
 * ```
 *  router = (new Router());
 *  ```
 */
```
or
```
/** \file
 * ```
 *  router = (new Router());
  * ```
 */
```

The relevant length is the length identifying the fenced code block / number of back ticks, not the number of spaces in front of it (analogous number of tildes).