QGDict::hashAsciiKey: Invalid null key due to empty procedure name
When having a problem like:
```
namespace eval ::tk::dialog {}
namespace eval ::tk::dialog::file {}

namespace eval ::tk::dialog::file::chooseDir {
    namespace import -force ::tk::msgcat::*
}

proc ::tk::dialog::file::chooseDir:: {args} {
}
```
This will lead to the following warnings:
```
QGDict::hashAsciiKey: Invalid null key
.../aa.tcl:9: warning: Illegal member name found.
```
this is due to the fact that the procedure name definition ends with `::` and thus actually has an empty procedure name (this is legal in TCL).
(see also: https://stackoverflow.com/questions/58683103/meaning-of-a-proc-name-ending-with).

In case of an empty name the last component is taken for the name too.