Spelling corrections for vhdl directory
Spelling corrections as found by codespell and in #561.

`assume_guarentee` is not a vhdl keyword, `assume_guarantee` is.
A number of spelling errors are in generated code (through javacc) and have been reported upstrean (https://github.com/javacc/javacc/pull/118)
Corection of `guarded_signal_specificatio` to `guarded_signal_specification` is not possible as `guarded_signal_specification` exists as well.