Better warning in case of `@form`
When having a problem like:
```
/** \file */

/**
 * The fie
 *
 * @form: the new format
 * @content: the content of the form
 *
 * Sets @form content to be @content.
 *
 * \f$ \alpha \f$ \f[ \beta \f]
 */
void fie(char *form, char *content);
```
we get warnings like:
```
.../aa.h:5: warning: Wrong formula id -1
.../aa.h:6: warning: Found unknown command '\content'
.../aa.h:8: warning: Wrong formula id -1
.../aa.h:8: warning: Found unknown command '\content'
```

The `@form:` is in RST analogue to the doxygen `@param form` but not understood by doxygen.
The result is that it is seen as an formula as doxygen translates the formula commands to `\form`.
For `@content` is handled properly as this is not a doxygen command.

In this patch a more unique name is chosen so these type of conflicts don't appear (that easy). The resulting, better, warnings are now:
```
.../aa.h:5: warning: Found unknown command '\form'
.../aa.h:6: warning: Found unknown command '\content'
.../aa.h:8: warning: Found unknown command '\form'
.../aa.h:8: warning: Found unknown command '\content'
```