Bug 163640 - Want INCLUDE_GRAPH_DEPTH configuration option
The request for a separate configuration option, from `MAX_DOT_GRAPH_DEPTH` and `DOT_GRAPH_MAX_NODES`, has not been fulfilled but based on the remarks about the functions / methods the commands `\callgraph` and `\callergraph` have been extended with the options `maxdepth` and `maxnodes`.
Analogous to the `\callgraph` `\hidecallgraph` the commands `\includegraph` `\hideincludegraph`, `\includedbygraph` and `\hideincludedbygraph` have been defined.

In case of usage multiple of the 'show' commands with the same name for an entity and confliction option values the maximum value is used (and a warning given).
In case the 'show' and 'hide' function are explicitly used for an entity also a warning is given.
In case an explicit command is given and the default is still valid (so non explicit command given) the explicit command values are taken.

The layout handling had to be extended slightly for the new commands as the entry regarding `includegraph` and `includedbygraph` not only is used to signal if a graph is visible but also the position in the output. Now we can have that generally no `includegrap` / `includedbygraph` is specified but that they are enabled by means of  `\includegraph` and `\includedbygraph`. in the old situation the layout element was not necessary but now we need the element to signal the right place (and hence need isVisible functionality also to set the proper default)