Correct style in table header in case of paragraph is used
In case a paragraph is used in a table header (quite unusual but can happen with 2 lines and an empty line in between), a `<p class="starttd">` tag is added and therewith the layout is incorrect.
This problem was detected based on #7409
This has been corrected (and tested with some available browsers FF, IE, Chrome, Opera).