Correction behavior of make clean in respect to version files
I case we ran `make clean` it was not possible anymore to generate doxygen as the generated file `generated_src/gitversion.cpp` was not present any more and it was not re-generated.
The generation of the `generated_src/gitversion.cpp` depends on teh `generated_src/git_state` file and this file was still present. This has been corrected.