issue #7558 PlantUML: Different behavior whether LATEX_OUTPUT has a final slash or not.
The problem is that full qualified paths in `*_OUTPUT` were not handled properly they were handled as it were relative paths.
In the documentation it is stated for e.g. LATEX_OUTPUT:

> The LATEX_OUTPUT tag is used to specify where the LATEX docs will be put. If a relative path is entered the value of OUTPUT_DIRECTORY will be put in front of it.

So in case of a non relative path the given path should be used and this path can be unrelated to `OUTPUT_DIRECTORY`, so we have to store the designated output path as well.