Problem in case using PREDEFINED with comma
In the pull request "Enable comma as separator in configuration lists enhancement " (#6563) it was made possible to have commas as separators for lists.
In case we have:
```
PREDEFINED             = A(x,y)=sin(x),cos(y)
```
and use `doxygen -x` (or usethe define): this results in:
```
PREDEFINED             = A(x \
                         y)=sin(x) \
                         cos(y)
```
this can be overcome by means of:
```
PREDEFINED             = "A(x,y)=sin(x),cos(y)"
```

But for a lot of existing packages this poses a problem.

(Found by looking at the doxygen configuration files as used by Fossies).