issue #7060 Function pointer typedefs render strangely in HTML documentation
typedefs like:
```
typedef void *(*pMsgCntrSubscriberAction)(const void* theContext);
```
place the `(*` part with the type, though it should be in front of the name., especially in HTML this gives a strange layout..
Splitting of this part and show it just before the name instead of with the type. The closing `)` has been split off and added to the arguments before.