issue #7626 "warning: documentation for unknown define ... found" without newline at end of file
In case of non preprocessed code there was already a `\n` added when not present at the end. This has now been done in case of preprocessing as well.

The current code had a small flaw as it added the `\n` after the last character and this character could be a `\0` so we has to backup a little.