issue #7406 Objective-C class and instance methods with the same name have the same refid
Methods in objectC can have a "+ for static methods and "-" for instance methods, this distinction was not taken into account.
In memberdef the distinction is made between both types so the correct anchor is created.
The changes in code.l are to get the right (member) definition in case of an objectC method (otherwise in e.g. the tooltip the reference went always to the first one found. This has also consequences for types.h, util.h and util.cpp as we need to add the information when searching for the right method.

Due to the incorporation of the "+ " and "- " also the test example has to be updated