Tcl has a problem with a case statement with just {} inside the switch statement
In case we have a (reduced) case like:
```
## ::tkcon::Init - inits tkcon
proc ::tkcon::Init {} {
            switch -glob -- $arg {
                -root           { set PRIV(root) $val }
                -font           { set OPT(font) $val }
                -rcfile {}
            }
}
```
and a default doxygen configuration file, doxygen will crash in the tcl scanner as the `{}` results in an empty string.

Some notes, crash only happens when compilation for 'Release; or 'RelWithDebInfo' as 'CMAKE_BUILD_TYPE'.

(Found by Fossies for magic 8.2.195).