Incorrect determination of fixed form Fortran
When we have the following small example, we see that the word `subroutine` doesn't start in column 7 but in column 8.
```
       subroutine expan()
c2345678  "          "     lb        "                "          "
       end
```
so the code is not converted from fixed form to free form Fortran and in the example (due to the odd number of double quoutes) result in:
```
********************************************************************
Error in file D:/speeltuin/bug_ftn_quote/small.f line: 5, state: 22(String)
********************************************************************
```

The condition regarding the column number was to restrictive.