Treat relative markdown links as a subpage instead of reference
It is a common case when project documentation is oranized as a set of markdown
files organized in a tree. Doxygen treats by default every markdown document as a
page and adds it into the root of tree view.

In existing implementation it may be worked around by adding Doxygen @subpage commands
in each markdown file and hide from Doxygen original reference. This results in a
cluttered markdown and upfron knowledge how ID for the markdown file is generated.

Example:

```markdown
Root.md
<!--
@cond HIDE_ME_FROM_DOXYGEN
-->

[Nested Document](nested/doc.md)

<!--
@endcond
-->

@subpage md_nested_doc
```

It would be much easier, if subpage relationship is handled by Doxygen itself also as resolution
of file name to reference ID.

This change is an adaptation of https://github.com/doxygen/doxygen/pull/7032/commits/e52b821579e52070d9f6ebd269850f30c8c1f807
with following changes:

- Minor fix to use `Portable` namespace.
- link is treated as @ref for absolute links and URLs
- link is treated as @subpage for relative links, file name resolved to ID

Signed-off-by: Andrey Vostrikov <andrey.vostrikov@cogentembedded.com>