issue #7678 Not all functions are created
Fenced code blocks with tildes and backticks were not handled in the same way in the respect that when after the backticks a language was given the complete sequence was handled by another rule.
By adding the backtick to the rule  where also the tilde was excluded fixes the problem.

This was also the case in the original problem, but was hidden as the `WARNINGS=NO` was set and thus messages like:
```
warning: reached end of file while inside a '```' block!
The command that should end the block seems to be missing!
```
where not shown.