Relative markdown file reference (#7032)
In case we have a relative reference to a local markdown file this file is not found when the relative path starts e.g. with `..`
So when we have a file:
```
docs\tutorial\security.md
```
and this references the file:
```
../api/browser-window.md
```
through the markdown syntax:
```
[`BrowserWindow`](../api/browser-window.md)
```
then the link was not found and a warning was given.