Text show as formula when USE_MATHJAX=YES
When having a line of code like:
```
callback_check = re.compile(r'([^\(]*\(.*)(\* *)(\).*\(.*\))')
```
this is seen as an incomplete formula when using MathJax, the `\(` is seen as start of a MathJax  formula.
Replacing the backslash by the corresponding code `&#92;` didn't work as this is already translated by the bowser and still picked up by MathJax, so we need `&zwj;` to separate the backslash and the bracket without any spacing.