Undefined control sequence for formula using MathJax
When having a code comment like:
```
@code
sub postprocess
{
    s/\(?\@xref\{(?:[^\}]*)\}(?:[^.<]|(?:<[^<>]*>))*\.\)?//g;
    s/\s+\(\@pxref\{(?:[^\}]*)\}\)//g;
}
@endcode
```

and using MathJax the `/(` is seen as a the beginning of a formula, but should be seen as text and can now lead to "Undefined control sequence".
(The problem comes originally from code as generated by the doxygen-perl-filter for converting perl code in something doxygen does understand).
This problem is similar to the problems solved in pull request #7697, but not handled here.
This pull request will see `\(` (and its counter part `\)` as complete entities and replace them in a similar way as done in #7697 (but now in an earlier stage but the `&zwj;` will be properly filtered in the different output formats.