issue #7734 Incorrect parsing of Q_PROPERTY
The parsing of the type / name was not done correctly as the name `{ID}` is also part of the `{TSCOPE}` and hence the name was seen as a type.
The name is the last part before an attribute is present.
Missing other parts:
- parsing of `*`
- not parsing of a number of (not supported attributes)

The definition of `Q_PROPERTY` is:
```
Q_PROPERTY(type name
           (READ getFunction [WRITE setFunction] |
            MEMBER memberName [(READ getFunction | WRITE setFunction)])
           [RESET resetFunction]
           [NOTIFY notifySignal]
           [REVISION int]
           [DESIGNABLE bool]
           [SCRIPTABLE bool]
           [STORED bool]
           [USER bool]
           [CONSTANT]
           [FINAL])
```
Note: in the implementation we do not enforce the order of the attributes.