Description of in out parameters
When a parameter is input parameter and output parameter on can specify `\param[in,out]` and in the description (with the aid of e.g. `<br>` or a list) make the distinction between the in and the out part though the both parts stay together.
When specifying separate `\param[in]` and `\param[out]` one gets a message like:
```
arning: argument 'a2' from the argument list of fie0 has multiple @param documentation sections
```

This patch makes it possible to have either one `\param` or one `\param[in]` together with one `\param[out]`