issue #7778 Build fails with javacc 5.0
At the moment doxygen works with javaCC version 7.0.5 but no test is done for the version. The minimum requirement has been set to 7.0.5
javacc does not have an option to retrieve the version though when giving `--version` a bit more information is given, this information is filtered.
(an issue has been submitted to javacc to obtain the version information: https://github.com/javacc/javacc/issues/172).