Warning in case of multiple python documentation sections.
In special cases when having multiple documentation section for 1 item in python it is possible that they are concatenated in a wrong way and result in a warning.
Each documentation section should be seen as a separate section and be separated from other sections.

the example:
```
#####################################################################
# # Modify Install Stage  ############################################
#####################################################################

class install(_install):
    """Specialised python package installer.

    It does some required chown calls in addition to the usual stuff.
    """
```
gives a warning like:
```
warning: unexpected command endverbatim
```