Empty enum setting in configuration file
In case we have an empty setting in the doxygen configuration file, where an enum value is expected like:
```
OUTPUT_LANGUAGE =
```
we get the warning:
```
warning: argument '(null)' for option OUTPUT_LANGUAGE is not a valid enum value
Using the default: English!
```

The default value should immediately have been used.