Addon option --pool to the test run script for parallel  execution of tests
To make better use of the processors of a computer the possibility has been build to execute the tests in parallel, this is especially of an advantage for PDF tests.
To make use of this option the number of parallel tests `np` has to be specified by means of `--pool=np`, the default is 1.