LaTeX test were hanging for a.o. Hungarian on Windows
On Windows a number of tests were hanging when using a language like Hungarian, Russian.
The handling of the output of the output is now done differently for Windows and other OS
A small (but obvious) error has been corrected (failed_html -> failed_latex), has no effect on the output / results though.