issue #7778 Build fails with javacc 5.0 (#7779)
* issue #7778 Build fails with javacc 5.0

At the moment doxygen works with javaCC version 7.0.5 but no test is done for the version. The minimum requirement has been set to 7.0.5
javacc does not have an option to retrieve the version though when giving `--version` a bit more information is given, this information is filtered.
(an issue has been submitted to javacc to obtain the version information: https://github.com/javacc/javacc/issues/172).

* issue #7778 Build fails with javacc 5.0

Version detection did not work for `5.0` (only x.y.z was handled, not x.y)

* issue #7778 Build fails with javacc 5.0

Based on a review
- `--version` is not a valid option, but present already for future use but
- removed double `  if(JAVACC_EXECUTABLE)`
- Needed `"${JAVACC_TEMP_VERSION}"` as otherwise when `javacc` is found but cannot be executed a message like:
  ```
  CMake Error at cmake/FindJavacc.cmake:11 (string):
    string sub-command REGEX, mode MATCH needs at least 5 arguments total to
    command.
  Call Stack (most recent call first):
    vhdlparser/CMakeLists.txt:1 (find_package)
  ```
  would appear
- reformulated some status strings

* issue #7778 Build fails with javacc 5.0

the `.` was not correctly escaped.

* issue #7778 Build fails with javacc 5.0

Another one 7.0........ match, going for the suggested match in the cmake CheckCXXCompilerFlag

* issue #7778 Build fails with javacc 5.0

Accidently left debug statement.

* issue #7778 Build fails with javacc 5.0

Problem has been fixed upstream for version 7.0.7 and up, not using `--version` but `-version` as all command line arguments are starting with a single '-'.
Also it is just the bare version, so small extra test so it will work with all versions.