Problem with generating test output for Chinese
The log output generated for the Chinese language can contain some extended ASCII characters, this results in a message like:
```
UnicodeDecodeError: 'utf-8' codec can't decode byte 0xf4 in position 36101: invalid continuation byte
```
and a crash of the test application during the read of the file.
In case of a crash it is attempted to open the file with another decoding and redo the read.