More flexible and correct detection of direction of an argument
In the current version a line like:
```
uint8_t innInstances,       ///< [inn]Number of CIP node instances.
```
was seen as an input parameter.
Also not all combinations were checked (`[inout]` was OK but `[outin]` wasn't although `[in,out]` and `[out,in]` were working in the tokenizer.
Line up the detection of the direction between the tokenizer and the method `extractDirection`

Note: this problem was found when working on #7879.