Fix compiling with clang parser support and with Clang compiler, too
When compiling the code with the Clang compiler a new CMake option
`use_libc++` is available which is enabled by default. It determines if
the Clang compiler should use libc++ as its C++ standard library or uses
the systems default (e.g. libstdc++ on Linux).
Disabling that option might help when compiling with Clang and the
option `use_libclang` is enabled but the associated library was not
built with libc++ as its C++ standard library.

Additionally, the doxygen executable that is created when CMake option
`static_libclang` is enabled, no longer crashes.