Fix compiling with clang parser support and with Clang compiler, too (#7922)
* Fix compiling with clang parser support and with Clang compiler, too

When compiling the code with the Clang compiler a new CMake option
`use_libc++` is available which is enabled by default. It determines if
the Clang compiler should use libc++ as its C++ standard library or uses
the systems default (e.g. libstdc++ on Linux).
Disabling that option might help when compiling with Clang and the
option `use_libclang` is enabled but the associated library was not
built with libc++ as its C++ standard library.

Additionally, the doxygen executable that is created when CMake option
`static_libclang` is enabled, no longer crashes.

* Removing obsolete (and wrong) forcing of -std=c++11 for GCC

The entire Doxygen project already requires the compiler to run in C++14
mode. That setting is applied globally in the top-most `CMakeLists.txt`
file.
Therefore, forcing GCC to run in C++11 mode in `src/CMakeLists.txt` is
wrong and obsolete.