crossref citations are shown unconditionally
The crossref items in a bib file are unconditionally added to the list of used citations although the the citation to which the crossref belongs to is not used.
This has been fixed. The problem was seen in the CGAL output.

Note: a crossref in LaTeX is a bit different from the normal understanding of cross-reference, from https://tex.stackexchange.com/questions/401138/what-is-the-bibtex-crossref-field-used-for: "crossref can be used if you have multiple entries referring to the same proceeding, book or similar."