bug_786382 Unmatched single quote in C++ source breaks aliases
Based on the definition in the (informative) Annex A of the C++ draft 2020 standard (N4849, part [gram.lex]), the definitions have been made for the lexer.
Now integer and floating point constants wit a single quote are seen as numbers and not as part of character constants.

(also tested on CGAL)