Warning about end of list in brief description after alias `^^` replacement
With the alias:
```
ALIASES += "sbl_add_package_main_class{2}=\brief \1 ^^ \2"
```
and the comment:
```
\sbl_add_package_main_class{Defines, \details dihedrals}
```
we will get the replacement:
```
\brief Defines \ilinebr \details dihedrals
```
but this leads to a number of warnings like:
```
warning: End of list marker found without any preceding list items
```

As the end of the brief description is here `\ilinebr` this is replaced in the definition by `\ilinebr.` (so with a `.`). We first have to strip the aritficial newlines at (the beginning and) the end end of the brief description before adding the `.`.

(Found as side effect of https://stackexchange.com/filters/57710/doxygen).