Double "->" in warning in case of trailing return type
When having the input
```
/*!
 * \brief Performs some side effect
 * \param i1 first                                                                                                                   */
auto side_effect_after(int  i1, int i2) -> void;
```
we get the warning
```
warning: The following parameter of side_effect_after(int i1, int i2) -> -> void is not documented:
```
so a double "->".
The "->" has already been added in scanner.l and defargs.l so shouldn't be added here.

(also the return type gives a warning, but this is another issue #6442)..