No warning for unbalanced brackets in markdown
When we have the input files qq.md.
```
@page xxx0 yyy0
@}
Start of text

@page xxx1 yyy1
@{
Start of text
```
and qqh.h:
```
/**

@page xxxh0 yyyh0
@}
Start of texth

*/
/**

@page xxxh1 yyyh1
@{
Start of texth

*/
```
we get the warnings:
```
qq.md:3: warning: unbalanced grouping commands
qqh.h:4: warning: unbalanced grouping commands
qqh.h:15: warning: end of file with unbalanced grouping commands
```
```
so we are missing
```
qq.md:9: warning: end of file with unbalanced grouping commands
```
due to the fact that the closing routine was not called (also the open routine was not called, always good to call it the set variables to their proper values.