Flush warnings file
When having a large project the warnings file is buffered but when a long time there are no warnings emitted the last warnings are not yet written.
During testing tis is most obvious when the `dot` process starts and one needs the `.dot` files but not the converted files at hat moment one wants to kill the doxygen process but the result is an incomplete warnings file of the warnings that should already be written.

Created a flush function for the warnings file and placed it at a few strategic points.