Incorrect recognition as RCS tag
When we have a, stripped down, example like:
```
\mainpage the page

<tt>$PATH:</tt>$x86$<tt>bin</tt>
```
we get the warning like:
```
aa.md:4: warning: end of comment block while expecting command </tt>
```
this is due to the fact that `$PATH:</tt>$` is seen as a RCS tag although `PATH` is not a RCS keyword.

Limiting the number of allowed keywords to the real RCS tags (for CVS the list is the same).

See also:
- https://www.gnu.org/software/trans-coord/manual/cvs/cvs.html#Keyword-substitution
- https://www.gnu.org/software/rcs/manual/rcs.html#Keywords