issue #8093 Some URLs in fresh/updated Doxyfile are split over two lines
Created an explicit split of the line before the http address so the URL will start on a new line in the doxygen configuration file and won't be split anymore (unless the total length of the  URL will be longer than 78 characters but this is in the doxygen settings not the case).
(The other formats: HTML, LateX and the help in the doxywizard give still the same output results).