Miscounting in case of markdown links
When we have simple input files like:
```
Chrome's trace viewer (Catapult) following the
[instructions](
https://chromium.googlesource.com/catapult/+/refs/heads/master/tracing/docs/embedding-trace-viewer.md). This directory contains
the helper files to embed Chrome's trace viewer.

The current resources were generated/copied from
[`Catapult@9508452e18f130c98499cb4c4f1e1efaedee8962`
](
https://chromium.googlesource.com/catapult/+/9508452e18f130c98499cb4c4f1e1efaedee8962).

\aa11
```
or
```
  Add new exported module Text.Pandoc.Writers.AnnotatedTable [API change]
  (#6655, Christian Despres).  This module (which should generally

  \bb3
```
one gets (in the current master version, a5ac10888f6c0b0fc00678f67aa49036b33aa4cf):
```
.../aa.md:8: warning: Found unknown command '\aa11'
.../bb.md:3: warning: Found unknown command '\bb4'
```
instead of
```
.../aa.md:11: warning: Found unknown command '\aa11'
.../bb.md:4: warning: Found unknown command '\bb4'
```
this is due to the fact that the newlines inside the links are not taken into consideration during the conversion.
We can add the extra newlines but have to do this inside the `<...>` as than they are handled correctly later on:
- when adding them before the `<` doxygen sees a newline and will start a new paragraph starting with the link instead of keeping it in its place.
- when adding them after the `>` we can get a warning about `warning: End of list marker found without any preceding list items` when after the closing `)` of the link there is directly a `.`