Incorrect handling of string with spaces and no quotes
During some tests on the documentation of LLVM the (CMake generated) doxygen configuration file contained the line:
```
DOT_PATH=D:\Program Files (x86)\Graphviz2.38\bin\dot.exe
```
and this was interpreted by doxygen as
```
DOT_PATH               = D:\ProgramFiles(x86)\Graphviz2.38\bin\dot.exe
```
without any message other  than that later on  the `dot` executable could not be found.
It is clear that here the double quotes were missing.

This problem has been solved by giving a warning and setting the value to the default value.
- configimpl.l
- config_doxyw.l

during the checking some debug facilities for the doxywizard were required and this has been implemented:
- config_doxyw.l
- doxywizard.h
- doxywizard.cpp

it was also discovered that in case of none existing int or bool values were specified the wrong "defaults" were taken (it was set to `0` / `false`), now the correct defaults are taken
- inputbool.cpp
- inputint.cpp