Handling commands for backslash and at-sign in ALIASES
When we have an command like `\\`, `\@`, `@@` or `@\`in ALIASES this should not disturb the recognition of the  replacement strings like `\2`, so e.g. `\\2` should not be replaced by `\y` but remain `\\2` (assuming  the `\2` is linked to `y` in an ALIASES).

This issue was found after #8137 had been fixed and was already present in older versions as well.