Incorrect checking between default and used values for strList in doxywizard
In case we use a default doxygen configuration file `doxygen -g` and use this as `doxywizard Doxyfile` we would expect not to see any differences between the used values and the default values, though we see in expert tab that `ABBREVIATE_BRIEF`, `STRIP_FROM_PATH` and `INPUT` don't have the default values  (item is "red"). This can also be observed in the run tab when selecting "Condensed" "Show configuration".
In the `isDefault()` not only the `strList` itself should be tested but also its values.